# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ContactInquiryCreate": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 10,
            "title": "Message",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "message"
        ],
        "title": "ContactInquiryCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Horoscope": {
        "properties": {
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "period": {
            "$ref": "#/components/schemas/HoroscopePeriod"
          },
          "sign": {
            "$ref": "#/components/schemas/ZodiacSign"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "sign",
          "period",
          "date",
          "content",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "Horoscope",
        "type": "object"
      },
      "HoroscopePeriod": {
        "enum": [
          "DAILY",
          "WEEKLY",
          "MONTHLY",
          "YEARLY"
        ],
        "title": "HoroscopePeriod",
        "type": "string"
      },
...
      "Page": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "seo_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Description"
          },
          "seo_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Title"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "title",
          "content",
          "id",
          "slug",
          "created_at",
          "updated_at"
        ],
        "title": "Page",
        "type": "object"
      },
...
      "Post": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "featured_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Featured Image"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus",
            "default": "DRAFT"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "title",
          "content",
          "slug",
          "id",
          "author_id",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "Post",
        "type": "object"
      },
...
      "PostListResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/Post"
            },
            "title": "Posts",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "posts"
        ],
        "title": "PostListResponse",
        "type": "object"
      },
...
      "PostStatus": {
        "enum": [
          "DRAFT",
          "PUBLISHED",
          "ARCHIVED"
        ],
        "title": "PostStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZodiacSign": {
        "enum": [
          "ARIES",
          "TAURUS",
          "GEMINI",
          "CANCER",
          "LEO",
          "VIRGO",
          "LIBRA",
          "SCORPIO",
          "SAGITTARIUS",
          "CAPRICORN",
          "AQUARIUS",
          "PISCES"
        ],
        "title": "ZodiacSign",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Aadikarta API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/public/contact": {
      "post": {
        "operationId": "submit_contact_inquiry_public_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInquiryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Contact Inquiry Public Contact Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Contact Inquiry",
        "tags": [
          "Public Content"
        ]
      }
    },
    "/public/horoscopes": {
      "get": {
        "operationId": "get_public_horoscopes_public_horoscopes_get",
        "parameters": [
          {
            "in": "query",
            "name": "sign",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ZodiacSign"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sign"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/HoroscopePeriod"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Horoscope"
                  },
                  "title": "Response Get Public Horoscopes Public Horoscopes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Horoscopes",
        "tags": [
          "Public Content"
        ]
      }
    },
    "/public/pages/{slug}": {
      "get": {
        "operationId": "get_public_page_public_pages__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Page",
        "tags": [
          "Public Content"
        ]
      }
    },
    "/public/posts": {
      "get": {
        "operationId": "get_public_posts_public_posts_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Posts",
        "tags": [
          "Public Content"
        ]
      }
    },
    "/public/posts/{slug}": {
      "get": {
        "operationId": "get_public_post_public_posts__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Post",
        "tags": [
          "Public Content"
        ]
      }
    },
...
  }
}